update to beta.13
