+++ node-cli/package.json
@@ -15,7 +15,7 @@
     "figlet": "^1.5.0",
     "inquirer": "^7.3.2",
     "ora": "^4.0.5",
-    "realm": "^10.0.0-beta.9"
+    "realm": "^10.0.0-beta.13"
   },
   "devDependencies": {
     "eslint": "^7.4.0",
