+++ node-cli/tasks.js
@@ -1,5 +1,4 @@
 const inquirer = require("inquirer");
-const Realm = require("realm");
 const bson = require("bson");
 const index = require("./index");
 const output = require("./output");
